+++ src/main/java/spring/backend/member/application/ReadMemberHomeService.java
@@ -9,7 +9,7 @@ import spring.backend.quickstart.dto.response.QuickStartResponse;
 import spring.backend.activity.query.dao.ActivityDao;
 import spring.backend.quickstart.query.dao.QuickStartDao;
 import spring.backend.member.domain.entity.Member;
-import spring.backend.member.dto.response.HomeMainResponse;
+import spring.backend.member.presentation.dto.response.HomeMainResponse;
 import spring.backend.member.dto.response.HomeMemberInfoResponse;
 
 import java.time.LocalDateTime;
