+++ src/main/java/spring/backend/member/presentation/dto/request/EditMemberProfileRequest.java
@@ -1,4 +1,4 @@
-package spring.backend.member.dto.request;
+package spring.backend.member.presentation.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
