+++ src/main/java/spring/backend/member/presentation/dto/response/HomeMainResponse.java
@@ -1,7 +1,8 @@
-package spring.backend.member.dto.response;
+package spring.backend.member.presentation.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import spring.backend.activity.dto.response.HomeActivityInfoResponse;
+import spring.backend.member.dto.response.HomeMemberInfoResponse;
 import spring.backend.quickstart.dto.response.QuickStartResponse;
 
 import java.util.List;
