+++ src/main/java/spring/backend/member/presentation/swagger/ReadMemberHomeSwagger.java
@@ -8,7 +8,7 @@ import spring.backend.core.configuration.swagger.ApiErrorCode;
 import spring.backend.core.exception.error.GlobalErrorCode;
 import spring.backend.core.presentation.RestResponse;
 import spring.backend.member.domain.entity.Member;
-import spring.backend.member.dto.response.HomeMainResponse;
+import spring.backend.member.presentation.dto.response.HomeMainResponse;
 import spring.backend.member.exception.MemberErrorCode;
 
 @Tag(name = "Member", description = "멤버")
